Improve performance of message translations
MessageFormat#format will look to replace the pattern with the arguments provided regardless of array length, however a lot of the translations in BungeeCord do not require an argument.
By skipping replacement if there are no arguments we can reduce object creation and remove unnecessary logic.

Adapted from WaterfallMC/Waterfall#39 with a micro-optimization

Signed-off-by: Janmm14 <computerjanimaus@yahoo.de>